# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Gestão de Usuários",
        "description": "API para gerenciamento de usuários e perfis, incluindo operações de criação, leitura, atualização e exclusão (CRUD).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Lista os perfis",
                "operationId": "43d12304c3e621ecce8cc9ccd4b3738e",
                "responses": {
                    "200": {
                        "description": "Lista de perfis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProfileResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Cria um perfil",
                "operationId": "00abb5f24a20112b161d17e9a5c38a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perfil criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profiles/{profile}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Busca um perfil",
                "operationId": "83d9b96185bae23c5b0a39be5b9681e7",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID do perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Atualiza um perfil",
                "operationId": "d287a4ee84dddce8a10e8dccba4c0ddf",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID do perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Exclui um perfil",
                "operationId": "5eb02443a95a4b74b56882493efb5017",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID do perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil excluído"
                    },
                    "404": {
                        "description": "Perfil não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProfileStoreRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Administrador",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "example": "Acesso administrativo",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProfileUpdateRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Administrador",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "example": "Acesso administrativo",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProfileResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "slug": {
                        "type": "string",
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Acesso administrativo",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-14T02:35:02.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-14T02:35:02.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token retornado por /login ou /register. Enviar como Authorization: Bearer {token}.",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}